Group isolation (#20)
* Sketch group types.

* Move party signup logic to session.

* Improve session signup logic.

* Add Group::new().

* Add group removal to client_disconnected().

* Add connection id to party signups.

Prepare session hashmap on Group.

* Prepare for group create and join.

* Draft logic for creating groups.

Include react router for routing.

* Debugging webpack worker bug.

* Using redux toolkit.

* Draft logic for joining a group.

Must wait for websocket connection to be opened before rendering.

* Hoist websocket code to UI thread.

To prevent multiple connections being opened due to the multiple worker
webpack bug.

It appears that Firefox has a bug (possibly a regression) that is also
opening multiple websocket connections.

* Draft logic for joining groups.

* Remove obsolete Parameters RPC call.

Threshold and parties are now chosen by the client when a group is
created.

* Draft session create server logic.

* Move client files to src directory.

* Create keygen session entry.

* Broadcast session create notifications.

* Using Keygen component.

* Improving keygen session handling.

* Draft logic for joining keygen sessions.

* Complete session join logic, prepare for signup.

* Keygen party signup.

* Begin migration to JSON RPC.

The ad-hoc message format had errors deserializing to the wrong type so
better to use an existing solution that works.

* Use RPC for group_join.

* Use RPC for session_create.

* Use RPC for session_join.

* Use RPC for session_signup.

* Restore session create broadcast notification.

* Comment out obsolete code.

* Fix multiple websocket connection bug.

* Fix broadcast notification logic.

So that broadcast notifications are constrained to the clients in a
group and optionally only those clients in a session.

* Remove obsolete CLI args.

Now that parties and threshold are associated with a group not with a
server instance.

* Broadcast SESSION_SIGNUP notification.

* Prepare comlink wrapped worker provider.

* Prepare client code for new state machine.

* Prepare server for peer relay.

* Fix server peer relay logic.

* Debug state transition timing issue.

* Debugging state transitions.

* Fix spurious notification warning.

Because during session_signup we need to wait for all parties then the
rpc_broadcast() method was warning on missing group_id before we reached
the limit as the notification context was not set yet.

Added a `noop` field to prevent this spurious warning.

* Debugging client bugs.

* Fix p2p answer handling.

There is a situation where all the p2p answers can be received before
the state transition has completed. Add the waitFor() function to ensure
that the state transition has exited and all the p2p answers are
available before moving on to the next transition.

* Tidy common module.

* Store the generated key share.

* Tidy onTransition handling.

* Migrate message signing state machine.

* Dispatch session finish notification.

So that we can be sure that all clients have completed generating their
key shares before automatically signing a message.

* Perform automatic signing on key generation.

* Generate public address from sign result.

* Tidy conditional.

* Use master branch from ZenGo-X/multi-party-ecdsa.

* Prepare to announce the public address.

Tidy RPC parameter handling for group create.

* Update json-rpc2 dependency.

* Tidy group/session handling in services.

* Send public address notification.

* Navigate to sign route.

* Tidy route handling.

* Update RPC method names.

* Update event names.

* Tidy group join handling.

* Prepare wallet connect form.

Draft sign() helper function.

* Only notify sessionCreate during keygen phase.

* Prepare session for signing proposal.

* Handle proposal notifications.

* WIP: Logic to sign proposals.

Currently not working, the signing is getting stuck :(

* WIP: fix party signup handling for sign.

* WIP: debugging sign issue.

* Remove obsolete types.

Debug notification events.

* Include round identifier in peer entry.

* Wait for sessionSignup event.

Fixes the bug with the sign state machine getting stuck.