Initial state persistence logic.
Exposes the thunks for clearState(), loadState() and saveState().

The state is an array of KeyShare instances which currently just
encapsulate a label but later will also include the key share
information generated using multi-party-ecdsa.

The data is encrypted and decrypted using xchacha20poly1305 functions
that are exposed via webassembly in the @metamask/mpc-snap-wasm package.

Currently, the webassembly is executed on the main UI thread as we are
dealing with small amounts of data it should execute fast enough.