Use keccak256 from ethers library.
This lets us clean up the WASM helpers a bit with a view to removing
them completely.

It is awkward because the ethers library does not give us a Uint8Array
but a 0x prefixed string which we then need to convert to a Uint8Array
and serialize as an Array.