Trying to set up fly.io.
Cannot publish to fly.io as the default `flyctl deploy` command times
out, see: https://github.com/superfly/flyctl/issues/1454.

Cannot publish a local build using --local-only as it is for the aarch64
architecture and fly.io requires x86_64.

Tried using docker buildx for a multi-architecture build but then the
container fails to build as the `cargo install` command fails in the
context of the container.